requirements.txt
kafka-python==2.0.2
psycopg[binary]==3.2.12